runpod-serverless/pyproject.toml
[project]
name = "runpod-serverless"
version = "0.1.0"
description = "A runpod serverless endpoint to generate image embeddings"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "runpod>=1.7.13",
    "torch>=2.8.0",
    "transformers>=4.56.1",
]
